[ISSUE-32] refactor: 랜덤 멍즈 데이터 처리 로직 utils로 분리 (#26)
